requirements.txt
langchain-openai
langchain-community
python-dotenv
ipykernel
bsy
langchain-core
streamlit